feat(mind): complete increment 7 addon consolidation and production gates
- Migrate study progress template to add-on JSONL bundle
- Move legacy graph ingest/project to addon graph_workflow layer
- Route insurance/hospital/property through adapter extract/project
- Add §13 production verification tests (disabled spy, offline ingest, confirm tokens)
- Fix characterization parity, memory session legacy import, and compile errors
