+++ packages/core_data/lib/src/storage/templates/template_registry.dart
@@ -6,9 +6,7 @@ import 'package:flutter/services.dart';
 class TemplateRegistry {
   TemplateRegistry._(this._templates);
 
-  static const bundledAssetPaths = <String>[
-    'packages/core_data/lib/src/storage/templates/study_progress.json',
-  ];
+  static const bundledAssetPaths = <String>[];
 
   final List<LifeTrackTemplate> _templates;
