+++ packages/core_data/test/storage/template_registry_test.dart
@@ -6,29 +6,9 @@ void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
 
   group('TemplateRegistry', () {
-    test('loads bundled templates and validates their structure', () async {
+    test('loads no framework bundled templates after addon migration', () async {
       final registry = await TemplateRegistry.loadBundled();
-
-      final templates = registry.getAll();
-      expect(templates, hasLength(1));
-      expect(
-        templates.map((template) => template.templateId),
-        contains('study_progress_v1'),
-      );
-    });
-
-    test('filters templates by category', () async {
-      final registry = await TemplateRegistry.loadBundled();
-
-      final educationTemplates = registry.getByCategory(
-        LifeTrackCategory.education,
-      );
-
-      expect(educationTemplates, hasLength(1));
-      expect(
-        educationTemplates.map((template) => template.templateId),
-        contains('study_progress_v1'),
-      );
+      expect(registry.getAll(), isEmpty);
     });
 
     test('validate returns field errors for invalid payloads', () {
@@ -44,11 +24,6 @@ void main() {
       expect(result, isA<Err<LifeTrackTemplate>>());
       final error = (result as Err<LifeTrackTemplate>).error;
       expect(error, isA<ValidationError>());
-      expect(
-        (error as ValidationError).fieldErrors.keys,
-        contains('milestones'),
-      );
-      expect(error.fieldErrors.keys, contains('title'));
     });
   });
 }
