+++ packages/feature_mind/addons/bundled_workflow_templates.jsonl
@@ -3,3 +3,4 @@
 {"template_id":"real_estate_under_construction_v1","title":"Under-Construction Flat Purchase Checklist","description":"7 critical verification steps before buying an under-construction flat","category":"realEstate","version":"1.0","milestones":[{"name":"Phase 1: RERA & Legal Verification","objective":"Verify builder legitimacy and project approvals","tasks":[{"summary":"Check Builder's RERA Registration","description":"Look into the builder's past delivery history and track record on the RERA website.","requirements":[{"label":"RERA Registration Number","type":"text","is_required":true,"hint":"Found on the official RERA portal"},{"label":"Builder Track Record Notes","type":"text","is_required":false,"hint":"Delayed projects count, pending projects count"},{"label":"RERA Verification Screenshot","type":"document","is_required":false}]},{"summary":"Verify Commencement Certificate Stage","description":"Check up to which floor the builder has approval and compare it with your target floor.","requirements":[{"label":"CC Approved Up To Floor","type":"number","is_required":true},{"label":"Your Target Floor","type":"number","is_required":true},{"label":"Commencement Certificate Copy","type":"document","is_required":false}]}]},{"name":"Phase 2: Financial Health Check","objective":"Assess the builder's and project's financial backing","tasks":[{"summary":"Review List of Approved Lenders","description":"Check which banks offer home loans for the project and treat an NBFC-only list as a warning sign.","requirements":[{"label":"Associated Banks List","type":"text","is_required":true,"hint":"List all banks offering loans for this project"},{"label":"Any NBFCs Only?","type":"boolean","is_required":true,"hint":"Red flag if only NBFCs, no major banks"}]},{"summary":"Check Land Encumbrance Certificate","description":"Verify whether the builder has taken a loan against the plot itself.","requirements":[{"label":"Encumbrance Certificate Document","type":"document","is_required":true},{"label":"Land Loan Status","type":"text","is_required":true,"hint":"Clear or outstanding, with amount if outstanding"},{"label":"Land Loan Clearance Date","type":"date","is_required":false}]}]},{"name":"Phase 3: Layout & Space Verification","objective":"Verify actual usable space matches what is promised","tasks":[{"summary":"Examine Carpet Area Layout","description":"Ask for structural layout drawings and note losses to ducts, lobbies, lifts, and columns.","requirements":[{"label":"Advertised Carpet Area (sq ft)","type":"number","is_required":true},{"label":"Structural Layout Drawing","type":"document","is_required":true,"hint":"Ask the builder explicitly for this"},{"label":"Actual Usable Area Notes","type":"text","is_required":false,"hint":"Space lost to lifts, ducts, columns, or lobbies"}]}]},{"name":"Phase 4: Legal Agreement Verification","objective":"Ensure all protections are in the buyer agreement","tasks":[{"summary":"Look for Escrow Account Clause","description":"Confirm the agreement explicitly references the mandated project escrow account.","requirements":[{"label":"Escrow Clause Present in Agreement?","type":"boolean","is_required":true},{"label":"Escrow Account Details","type":"text","is_required":false,"hint":"Bank name or account reference if available"},{"label":"Buyer Agreement Document","type":"document","is_required":false}]},{"summary":"Confirm Timeline for Promised Amenities","description":"Check the legal agreement for definitive delivery dates on every amenity shown in brochures.","requirements":[{"label":"Amenities Delivery Date in Contract?","type":"boolean","is_required":true},{"label":"Promised Amenities List","type":"text","is_required":true,"hint":"Clubhouse, gym, garden, pool, and similar amenities"},{"label":"Contract Delivery Date","type":"date","is_required":false},{"label":"Brochure vs Agreement Comparison Notes","type":"text","is_required":false}]}]}]}
 {"template_id":"car_purchase_v1","title":"Car Purchase Decision Template","description":"A bundled template for verifying prerequisites, researching vehicles, and planning financing before purchase.","category":"carPurchase","version":"1.0","milestones":[{"name":"Phase 1: Prerequisites & Documentation","objective":"Confirm license, fitness, and required personal documents","tasks":[{"summary":"Verify driving license validity","requirements":[{"label":"Driving License Copy","type":"document","is_required":true}]},{"summary":"Confirm physical fitness and parking readiness","requirements":[{"label":"Physical Fitness Notes","type":"text","is_required":false},{"label":"Parking Plan","type":"text","is_required":true}]}]},{"name":"Phase 2: Car Research & Selection","objective":"Compare shortlisted vehicles against budget and needs","tasks":[{"summary":"Compare shortlisted cars","requirements":[{"label":"Shortlisted Cars","type":"text","is_required":true},{"label":"Test Drive Notes","type":"text","is_required":false}]}]},{"name":"Phase 3: Financial Planning & Loan Pre-approval","objective":"Secure budget clarity, insurance estimates, and pre-approved financing","tasks":[{"summary":"Evaluate total ownership cost","requirements":[{"label":"Budget Range","type":"number","is_required":true},{"label":"Insurance Estimate","type":"number","is_required":false}]},{"summary":"Collect loan pre-approval options","requirements":[{"label":"Loan Offers","type":"text","is_required":true}]}]}]}
 {"template_id":"university_admission_v1","title":"University Admission Planning Template","description":"A bundled skeleton for shortlisting programs, preparing documents, applying, and completing enrollment.","category":"education","version":"1.0","milestones":[{"name":"Phase 1: Shortlisting","objective":"Identify target programs and deadlines","tasks":[{"summary":"Build university shortlist","description":"List target, reach, and safety programs with deadlines.","requirements":[{"label":"University Shortlist","type":"text","is_required":true}]}]},{"name":"Phase 2: Document Preparation","objective":"Prepare transcripts, test scores, essays, and recommendations","tasks":[{"summary":"Collect required application documents","requirements":[{"label":"Document Checklist","type":"text","is_required":true}]}]},{"name":"Phase 3: Applications","objective":"Submit applications and track results","tasks":[{"summary":"Submit applications before deadlines","requirements":[{"label":"Submitted Programs","type":"text","is_required":true}]}]},{"name":"Phase 4: Visa & Enrollment","objective":"Complete admission acceptance and relocation steps","tasks":[{"summary":"Prepare visa and enrollment confirmations","requirements":[{"label":"Visa or Enrollment Notes","type":"text","is_required":true}]}]}]}
+{"template_id":"study_progress_v1","title":"Study Progress Template","description":"A bundled skeleton for recording what you are studying, last topic, and the next session \u2014 stored locally as a LifeTrack.","category":"education","version":"1.0","milestones":[{"name":"Phase 1: Subject","objective":"Name the course or subject this journey tracks","tasks":[{"summary":"Record the subject","requirements":[{"label":"Subject","type":"text","is_required":true,"hint":"For example Java, DSA, or Physics"}]}]},{"name":"Phase 2: Current work","objective":"Keep the latest topic and a short progress note","tasks":[{"summary":"Log the current topic","requirements":[{"label":"Current Topic","type":"text","is_required":false},{"label":"Progress Notes","type":"text","is_required":true}]}]},{"name":"Phase 3: Next session","objective":"Capture the exam or goal date and what comes next","tasks":[{"summary":"Plan the next study block","requirements":[{"label":"Exam or Goal Date","type":"date","is_required":false},{"label":"Next Session Plan","type":"text","is_required":false}]}]}]}
