+++ packages/feature_mind/lib/src/addons/templates/addon_template_catalog.dart
@@ -40,6 +40,7 @@ class AddonTemplateCatalog {
     'packages/feature_mind/addons/property-purchase-planner/addon.json',
     'packages/feature_mind/addons/car-purchase-planner/addon.json',
     'packages/feature_mind/addons/university-admission-planner/addon.json',
+    'packages/feature_mind/addons/study-progress-tracker/addon.json',
   ];
 
   static Future<AddonTemplateCatalog> loadBundled({AssetBundle? bundle}) async {
