+++ packages/feature_mind/pubspec.yaml
@@ -94,6 +94,8 @@ flutter:
     - addons/car-purchase-planner/lifetrack_template.json
     - addons/university-admission-planner/addon.json
     - addons/university-admission-planner/lifetrack_template.json
+    - addons/study-progress-tracker/addon.json
+    - addons/study-progress-tracker/lifetrack_template.json
     - addons/draft-diet-plan/addon.json
   # No bundled `assets/models/` any more. Neither app ships model weights in
   # the APK -- both default to `DownloadModelProvider`
